SMHE-191: An external SIG analysis, recommended to make this not a public class. This class could be vulnerable to arbitrary command execution. The reason is that this class executes a command on the operating system.
